requirements.txt
langchain
langchain-openai
python-dotenv
numpy
scikit-learn
faiss-cpu
wandb
pypdf
unstructured
markdown
langchain-community
sentence-transformers
faiss-cpu
langchain_huggingface
